data_pipeline/requirements.txt
openai
anthropic
jsonlines
scrapfly-sdk
python-dotenv
lxml
tenacity
tqdm
ipykernel
regex
datasets